+++ scorer/aligner.py
