+++ scorer/conn_head_mapper.py
