+++ scorer/const.py
