+++ scorer/example_partial_scoring.pdf
