+++ scorer/partial scoring example.pdf
