+++ scorer/validator.py
